Moved PeriodicTableWidget into AlgorithmDialogs,re #23667
